+++ CHANGES.rst
@@ -7,6 +7,7 @@ in progress
 -----------
 - lqdn-gafam-campaign: Obtain "lang" and "name" variables from query parameters,
   e.g. https://examples.posterkit.net/lqdn-gafam-campaign/poster.html?lang=fr&name=google
+- lqdn-gafam-campaign: Apply custom layout settings based on poster name
 
 2018-01-15 0.0.0
 ----------------